Essentials for UI update
New contrains have been added
New gauges -> power gauge is seperate now
Joystick has been modified
Placeholders for time & battery

Refs #61

+++ client/app/src/main/java/group01/smartcar/client/SmartCar.java
@@ -47,6 +47,6 @@ public interface SmartCar {
     }
 
     interface MotorPowerUpdatedCallback {
-        void onMotorPowerUpdated(double power);
+        void onMotorPowerUpdated(int power);
     }
 }
